Implemented mzal's suggestions and fixed bugs
Indentation fixed.
Getting colour numbers now pulled out into separate function.
Getting colour numbers now actually works.
Removed all 'int's from GPU.
Added signed contextless types: s8 and s16.
Shortened _put_ functions throught ternaries.
Getting tile number and attributes now pulled out into separate
function.

+++ src/include/gpu.h
@@ -19,7 +19,7 @@ void gpu_write_spd(d8 new_spd);
 
 
 void gpu_prepare(char * rom_title);
-void gpu_step(int cycles_delta);
+void gpu_step(s8 cycles_delta);
 void gpu_destroy(void);
 
 #endif /* GPU_H_ */
